% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{DD2Rad}
\alias{DD2Rad}
\title{Degree to Radian Converter}
\usage{
DD2Rad(phi_deg)
}
\arguments{
\item{phi_deg}{Optional. A numeric scalar that denotes the latitude in terms of degree [degree].}
}
\value{
The function convert the value of an angel in the unit degree to the unit radian as a numeric scalar.
}
\description{
\code{DD2Rad} converts the value of an angel in the unit degree to the unit radian.
}
\details{
This is a function to convert the degree unit to radian.
}
\section{Reference}{

\href{https://www.fao.org/3/x0490E/x0490e00.htm}{Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998).
   \emph{Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56}.
   Fao, Rome, 300(9), D05109.}
}

\examples{
DD2Rad(phi_deg = 60.73)

}
\seealso{
\code{\link{SunHA}}.
}
