% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{EarSunDis}
\alias{EarSunDis}
\title{Inverse Earth-Sun Distance (d_r)}
\usage{
EarSunDis(date)
}
\arguments{
\item{date}{Optional. A character string that denotes the date in the format "Year-Month-Day" or "Year/Month/Day".}
}
\value{
The function returns the value of inverse relative earth-sun distance as a numeric scalar.
}
\description{
\code{EarSunDis} returns the inverse earth-sun distance.
}
\details{
This is a function to calculate the inverse earth-sun distance.
}
\section{Reference}{

\href{https://www.fao.org/3/x0490E/x0490e00.htm}{Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998).
   \emph{Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56}.
   Fao, Rome, 300(9), D05109.}
}

\examples{
EarSunDis("2020/08/25")

}
\seealso{
\code{\link{JulDate}}.
}
