% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffPrec.R
\name{EffPrec}
\alias{EffPrec}
\title{Effective Monthly Precipitation (P_eff)}
\usage{
EffPrec(P_tot)
}
\arguments{
\item{P_tot}{A numeric scalar that denotes the total monthly precipitation [mm].}
}
\value{
The function returns the value of effective monthly precipitation [mm].
}
\description{
\code{EffPrec} returns the value of effective precipitation.
}
\details{
This is a function to calculate the effective precipitation [mm]. The function formula has been developed
   for Iran where the mean annual precipitation is about 250 mm. It may be used for similar semi-arid areas,
   but it is not recommended for the areas with different climate.
}
\examples{
EffPrec(P_tot = 450)

}
