% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO56.R
\docType{package}
\name{FAO56-package}
\alias{FAO56-package}
\title{A package for computing the crop evapotranspiration and evapotranspiration rate
   from the reference surface by FAO Penman-Monteith equation}
\description{
FAO56 provides the equations used to calulate the crop evapotranspiration and evapotranspiration rate
   from the reference surface by FAO Penman-Monteith equation based on \emph{FAO paper No, 56:
   Crop evapotranspiration - Guidelines for computing crop water requirements}
}
