% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_botanical_names}
\alias{fao_botanical_names}
\title{FAO botanical names}
\format{
A \code{tibble} containing the crop coefficients
}
\usage{
fao_botanical_names
}
\description{
FAO botanical names.
}
\keyword{datasets}
