% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_crop_classes}
\alias{fao_crop_classes}
\title{FAO crop classifications}
\format{
A \code{tibble} containing the crop classifications
}
\usage{
fao_crop_classes
}
\description{
FAO crop classifications.
}
\keyword{datasets}
