% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_crop_coefficients}
\alias{fao_crop_coefficients}
\title{FAO crop coeffients}
\format{A \code{tibble} containing the crop coefficients}
\usage{
fao_crop_coefficients
}
\description{
FAO crop coefficients from FAO 56.
}
\keyword{datasets}
