% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fao_crop_stages}
\alias{fao_crop_stages}
\title{FAO crop growth stages}
\format{A \code{tibble} containing the crop coefficients}
\usage{
fao_crop_stages
}
\description{
FAO crop growth stages from FAO 56.
}
\keyword{datasets}
