% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{gen_ETo_predictors_from_aquastat}
\alias{gen_ETo_predictors_from_aquastat}
\title{Generate ETo predictors from aquastat data}
\usage{
gen_ETo_predictors_from_aquastat(clim_data, lat, z, col_select = NULL)
}
\arguments{
\item{clim_data}{\code{data.frame} of monthly climate data in order of Jan-Dec, containing the columns described below}

\item{lat}{latitude in decimal degrees}

\item{z}{elevation in m}

\item{col_select}{character vector specifying which columns to return from the original data}
}
\description{
Generate ETo predictors from aquastat data
}
\details{
\itemize{
\item Rel_Hum_pct
\item Tmp_max_degC
\item Tmp_min_degC
\item Tmp_Mean_degC
\item Sun_shine_pct (0-100)
\item Wind_2m_m_per_s
}
}
\examples{
Locate and read the example et0 csv file
et0_path <- system.file("extdata", "ET0_example_file.csv", package = "fao56")
col_select <- c("Rn_MJ_per_day", "G_MJ_per_day", "Tmp_Mean_degC", "Wind_2m_m_per_s", "es_kPa", "ea_kPa", "ETo_calc", "ETo_mm_per_d")
clim_data <- read_et0_csv(et0_path)
ETo_df <- gen_ETo_predictors_from_aquastat(clim_data, lat = 20.59, z = 231, col_select = col_select)
}
