% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_aquastat_climate.R
\name{get_aquastat_climate}
\alias{get_aquastat_climate}
\title{Get Aquastat Climate}
\usage{
get_aquastat_climate(lat, lon)
}
\arguments{
\item{lat}{latitude of location}

\item{lon}{longitude of location}
}
\description{
This function takes you to the download pages for AQUASTAT CLIMATE
}
\details{
This function opens a webpage to download data at the lat/lon specified
}
