% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_crop_classification.R
\name{get_crop_classes}
\alias{get_crop_classes}
\title{Get FAO Crop Classifications}
\usage{
get_crop_classes()
}
\value{
A list containing two \code{data.frame} objects: \code{crop_classes} and \code{botanical_names}.
}
\description{
Get FAO Crop Classifications and botanical names
}
\examples{
crop_classification <- get_crop_classes()
crop_classes <- crop_classification$crop_classes
botanical_names <- crop_classification$botanical_names
}
