% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_crop_tables.R
\name{get_crop_tables}
\alias{get_crop_tables}
\title{Get FAO 56 Tables}
\usage{
get_crop_tables()
}
\value{
A list containing two \code{data.frame} objects: \code{crop_stages} and \code{crop_coeff}.
}
\description{
Get FAO Crop Coefficients and growing periods
}
\examples{
fao_tables <- get_crop_tables()
crop_stages <- fao_tables$crop_stages
crop_coeff <- fao_tables$crop_coeff
}
