% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_d_r}
\alias{get_d_r}
\title{Get inverse relative distance Earth-Sun (Eq 23)}
\usage{
get_d_r(date)
}
\arguments{
\item{date}{date object or numeric jday}
}
\description{
Get inverse relative distance Earth-Sun (Eq 23)
}
\details{
Return inverse relative distance. Evaluates the equation:

d_r = 1 + 0.033 * cos(2 * pi/ 365 * jday)
}
\examples{
get_d_r("2019-04-15")
get_d_r(105, "\%j")
}
