% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_delta}
\alias{get_delta}
\title{Get solar decimation (Eq 24)}
\usage{
get_delta(date)
}
\arguments{
\item{date}{date object or numeric jday}
}
\description{
Get solar decimation (Eq 24)
}
\details{
Return solar decimation in rad. Evaluates the equation:

delta = 0.409 * sin(2*pi/365 * jday - 1.39)
}
\examples{
\dontrun{
get_delta("2019-04-15")
get_delta(105, "\%j")
}
}
\keyword{internal}
