% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_ea_from_RHmean}
\alias{get_ea_from_RHmean}
\title{Get vapor pressure from RHmean and temperature}
\usage{
get_ea_from_RHmean(RHmean, Tmax_C, Tmin_C)
}
\arguments{
\item{RHmean}{Mean relative humidity, pct (0-100)}

\item{Tmax_C}{Max daily temperature, deg C}

\item{Tmin_C}{Min daily temperature, deg C}
}
\description{
Get vapor pressure from RHmean and temperature
}
\details{
Calculate the actual vapor pressure from mean relative humidity and
  max and min daily temperature. The calculation is as:

  ea = RHmean / 100 * (es(Tmax_C) + es(Tmin_C)) / 2 (FAO 56 Eq. 19)

  es(T) is the saturation vapor pressure at that temperature. The function
  \code{get_es(T_C)} is used for saturation vapor pressure.
}
\examples{
get_ea_from_RHmean(RHmean = 68, Tmax_C = 25, Tmin_C = 18)
}
