% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_es}
\alias{get_es}
\title{Calculate the saturation vapor pressure in kPa}
\usage{
get_es(T_C)
}
\arguments{
\item{T_C}{Temperature in deg C}
}
\description{
Calculate the saturation vapor pressure in kPa
}
\details{
Calculates the saturation vapor pressure in kPa as a function of temperature (deg C) as

es = a * exp(b * T_C / (T_C + c1)), where

a = 0.6108
b = 17.27
c = 237.3
}
\examples{
get_es(24.5)
}
