% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_es_slope}
\alias{get_es_slope}
\title{Calculate the slope of the saturation vapor pressure curve
culate the slope of the saturation vapor pressure curve in kPa / deg C}
\usage{
get_es_slope(T_C)
}
\arguments{
\item{T_C}{Temperature in deg C}
}
\description{
Calculate the slope of the saturation vapor pressure curve
culate the slope of the saturation vapor pressure curve in kPa / deg C
}
\details{
Calculates the saturation vapor pressure in kPa / deg C as a function of temperature (deg C) as

s = a * b * c / (T_C + c)^2 * exp(b * T_C / (T_C + c)), where

a = 0.6108
b = 17.27
c = 237.3
}
\examples{
get_es_slope(24.5)
}
