% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_P_atmosphere}
\alias{get_P_atmosphere}
\title{Get atmospheric pressure at elevation}
\usage{
get_P_atmosphere(z)
}
\arguments{
\item{z}{elevation, masl}
}
\description{
Get atmospheric pressure at elevation
}
\details{
Get atmospheric pressure, in kPa, at elevation, in m
}
\examples{
\dontrun{
get_P_atmosphere(0)
}
}
