% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_phi_lat}
\alias{get_phi_lat}
\title{Get latitude in rad from decimal degrees (Eq. 22)}
\usage{
get_phi_lat(lat)
}
\arguments{
\item{lat}{latitude in decimal degrees}
}
\description{
Get latitude in rad from decimal degrees (Eq. 22)
}
\details{
Return latitude in rad. Evaluates the equation

lat * pi / 180
}
\examples{
\dontrun{
phi <- get_phi_lat(13.1)
}
}
\keyword{internal}
