% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_Ra_daily}
\alias{get_Ra_daily}
\title{Extraterrestrial radiation for daily periods, Ra (MJ/m^2/day)}
\usage{
get_Ra_daily(lat, date)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{date}{date object or numeric jday}
}
\description{
Get daily extraterrestrial radiation (FAO 56, Ch 3, Eq. 21)
}
\details{
Returns daily extraterrestrial radiation in units of MJ/m^2/day
per FAO 56, Ch 3, Eq. 21. Evaluates the equation

Ra <- 24 * 60 / pi * Gsc * d_r *
  (w_s * sin(phi) * sin(delta) + cos(phi) * cos(delta) * sin(w_s))

To do this, it evaluates internal functions to obtain
Gsc solar constant 0.0820, MJ/m^2/day
d_r inverse relative distance Earth-Sun (Equation 23)
w_s sunset hour angle (Equation 25 or 26), rad,
phi latitude (Equation 22), rad
delta solar decimation (Equation 24), rad.
}
\examples{
lat <- -22.9 # Rio de Janeiro
date <- "2019-05-15"
get_Ra_daily(lat, date)
}
