% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_Rnl_daily}
\alias{get_Rnl_daily}
\title{Get net outgoing longwave radiation (Rnl)}
\usage{
get_Rnl_daily(Ra, Tmax_K, Tmin_K, ea, Rs, Rso)
}
\arguments{
\item{Ra}{downward shortwave radiation, MJ/m^2/day}

\item{Tmax_K}{maximum absolute temperature during the 24-hour period, K = deg
C + 273.16}

\item{Tmin_K, }{K minimum absolute temperature during the 24-hour period, K =
deg C + 273.16}

\item{ea}{actual vapour pressure, kPa}

\item{Rs}{measured or calculated. (Equation 35) solar radiation [MJ m-2
day-1],}

\item{Rso}{calculated (Equation 36 or 37) clear-sky radiation [MJ m-2 day-1].}
}
\description{
Get net outgoing longwave radiation, Rnl (FAO 56, Eq. 39) in MJ/m^2/day
}
\details{
This function calculates net outgoing longwave radiation, in
  MJ/m^2/day. It uses Eq. 39 from FAO 56 in Chapter 3:

  Rnl = sigma * (Tmax_K + Tmin_K)/2 * (0.34 - 0.14 * sqrt(e_a)) * (1.35 *
  Rs/Rso - 0.35)


  sigma is the Stefan-Boltzmann constant, 4.903e-9 MJ/K^4/m^2/day,
  Rs/Rso relative shortwave radiation (limited to <= 1.0).
}
\examples{
lat <- -22.9 # Rio de Janeiro
date <- "2019-05-15"
n <- 220 / 31 # 220 hours in a month / 31 days
N <- get_daylight_hours(lat, date)
Ra <- get_Ra_daily(lat, date)
Rso <- get_Rso_daily(Ra, z = 100)
Rs <- get_Rs_daily(Ra, n, N)
Tmax_K <- 25.1+273.16
Tmin_K <- 19.1+273.16
ea <- 2.1
Rnl <- get_Rnl_daily(Ra, Tmax_K, Tmin_K, ea, Rs, Rso)
}
