% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_Rns_daily}
\alias{get_Rns_daily}
\title{Get net shortwave radiation (Rs)}
\usage{
get_Rns_daily(Rs, albedo = 0.23)
}
\arguments{
\item{Rs}{downward shortwave radiation, MJ/m^2/day}

\item{albedo}{albedo or canopy reflectance, unitless}
}
\description{
Get net shortwave radiation, Rs (FAO 56, Eq. 38) in MJ/m^2/day
}
\details{
This function calculates net shortwave radiation, in MJ/m^2/day. It
  uses Eq. 38 from FAO 56 in Chapter 3:

  Rns = (1 - albedo) * Rs

  Note that for reference ET0, albedo can be taken as 0.23 for hypothetical
  grass reference crop.
}
\examples{
lat <- -22.9 # Rio de Janeiro
date <- "2019-05-15"
n <- 220 / 31 # 220 hours in a month / 31 days
N <- get_daylight_hours(lat, date)
Ra <- get_Ra_daily(lat, date)
Rs <- get_Rs_daily(Ra, n, N)
Rns <- get_Rns_daily(Rs, albedo = 0.23)
}
