% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_Rs_daily}
\alias{get_Rs_daily}
\title{Get solar radiation (Rs)}
\usage{
get_Rs_daily(Ra, n, N, as = 0.25, bs = 0.5)
}
\arguments{
\item{Ra}{extraterrestrial radiation, MJ/m^2/day}

\item{n}{actual duration of sunshine, hour}

\item{N}{maximum possible duration of sunshine or daylight hours, hour}

\item{as}{regression constant, expressing the fraction of extraterrestrial
radiation reaching the earth on overcast days (n = 0)}

\item{bs}{increase in solar radiation on clear days relative to overcast
days, ie, as + bs is the fraction of extraterrestrial radiation reaching
the earth on clear days (n = N).}
}
\description{
Get solar radiation, Rs (FAO 56, Eq. 35) in MJ/m^2/day
}
\details{
This function calculates solar or shortwave radiation, in
  MJ/m^2/day. It uses Eq. 35 from FAO 56 in Chapter 3:

  Rs = (as + bs *  n / N) * Ra

  Note that \code{n/N} is the relative sunshine duration (unitless). FAO56
  suggests using as = 0.25 and bs = 0.5 if observed values are not available.
  But better to use observations if possible.
}
\examples{
lat <- -22.9 # Rio de Janeiro
date <- "2019-05-15"
n <- 220 / 31 # 220 hours in a month / 31 days
N <- get_daylight_hours(lat, date)
Ra <- get_Ra_daily(lat, date)
Rs <- get_Rs_daily(Ra, n, N)
}
