% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_Rso_daily}
\alias{get_Rso_daily}
\title{Get clear sky shortwave radiation (Rso)}
\usage{
get_Rso_daily(Ra, z)
}
\arguments{
\item{Ra}{downward shortwave radiation, MJ/m^2/day}

\item{z}{elevation above sea level, m}
}
\description{
Get clear sky shortwave radiation, Rso (FAO 56, Eq. 36) in MJ/m^2/day
}
\details{
This function calculates clear-sky shortwave radiation, in MJ/m^2/day. It
  uses Eq. 37 from FAO 56 in Chapter 3:

  Rso = (0.75 + 2e-5*z) * Ra

  Note that different values for as and bs should be used if available and
  then a different equation (36) can be used.
}
\examples{
lat <- -22.9 # Rio de Janeiro
date <- "2019-05-15"
n <- 220 / 31 # 220 hours in a month / 31 days
N <- get_daylight_hours(lat, date)
Ra <- get_Ra_daily(lat, date)
Rso <- get_Rso_daily(Ra, z = 100)
}
