% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{ggsub}
\alias{ggsub}
\title{Pipe-friendly gsub}
\usage{
ggsub(x, pattern, replacement)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can
be coerced by as.character to a character vector. Long vectors are
supported.}

\item{pattern}{character string containing a regular expression (or character
string for fixed = TRUE) to be matched in the given character vector.
Coerced by as.character to a character string if possible. If a character
vector of length 2 or more is supplied, the first element is used with a
warning. Missing values are allowed except for regexpr, gregexpr and
regexec.}

\item{replacement}{a replacement for matched pattern in sub and gsub. Coerced
to character if possible. If a character vector of length 2 or more is
supplied, the first element is used with a warning. If NA, all elements in
the result corresponding to matches will be set to NA.}
}
\description{
Pipe-friendly gsub
}
