% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Kc_Vegetables_Cucumber_Family}
\alias{Kc_Vegetables_Cucumber_Family}
\title{Crop Coefficients (Kc) of Cucumber Family Vegetables}
\format{
A data frame with 7 rows and 4 variables:
\describe{
  \item{Crop}{name of the crop}
  \item{Kc_ini}{the crop coefficient in the growth initial stage}
  \item{Kc_mid}{the crop coefficient in the growth mid-season stage}
  \item{Kc_end}{the crop coefficient in the growth late-season stage}
}
}
\source{
\url{https://www.fao.org/3/x0490E/x0490e00.htm}
}
\description{
A dataset containing the crop coefficients (Kc) of the cucumber family vegetables extracted from Table 12 in
"Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56".
It is a source to assign a suitable value to the argument \code{Kc} of the function \code{ET_c}.
}
