% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_CLIMWAT.R
\name{read_CLIMWAT}
\alias{read_CLIMWAT}
\title{Read CLIMWAT data}
\usage{
read_CLIMWAT(directory_path, station_name)
}
\arguments{
\item{directory_path}{Path to directory containing CLIMWAT files}

\item{station_name}{Name of station to import (will import .cli and .pen files)}
}
\description{
Import CLIMWAT data from .cli and .pen files. Note that the files must
first be extracted from CLIMWAT and saved as .cli and .pen files.
}
\examples{
\dontrun{
directory_path <- "fao56"
station_name <- "CLIMWAT_example"
read_CLIMWAT(directory_path,station_name)
}
}
