% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radiation.R
\name{SolDec}
\alias{SolDec}
\title{Solar Declination (delta)}
\usage{
SolDec(date)
}
\arguments{
\item{date}{Optional. A character string that denotes the date in the format "Year-Month-Day" or "Year/Month/Day".}
}
\value{
The function returns the value of solar declination as a numeric scalar.
}
\description{
\code{SolDec} returns the solar declination.
}
\details{
This is a function to calculate the value of solar declination.
}
\section{Reference}{

\href{https://www.fao.org/3/x0490E/x0490e00.htm}{Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998).
   \emph{Crop evapotranspiration - Guidelines for computing crop water requirements - FAO Irrigation and drainage paper 56}.
   Fao, Rome, 300(9), D05109.}
}

\examples{
SolDec("2020/08/25")

}
\seealso{
\code{\link{JulDate}, \link{SunHA}}.
}
