% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_psychrometric_constant}
\alias{get_psychrometric_constant}
\title{Calculate the psychrometric constant}
\usage{
get_psychrometric_constant(P_kPa = 101.325)
}
\arguments{
\item{P_kPa}{Pressure in kPa}
}
\description{
Calculate the psychrometric constant
}
\details{
Calculates the psychrometric constant as

gamma = cp * P / (M_water_air * l), org

gamma = 0.663e-3 P

gamma psychrometric constant, kPa / deg C,
P atmospheric pressure, kPa
l latent heat of vaporization, 2.45 MJ / kg,
cp specific heat at constant pressure, 1.013e-3, MJ / kg / deg C
M_water_air ratio molecular weight of water vapor/dry air = 0.622.
}
\examples{
get_psychrometric_constant(101.325)
}
