% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{get_w_s_angle}
\alias{get_w_s_angle}
\title{Get sunset hour angle (Eq 25)}
\usage{
get_w_s_angle(lat, date)
}
\arguments{
\item{phi}{latitude (Equation 22), rad}

\item{delta}{solar decimation (Equation 24), rad.}
}
\description{
Get sunset hour angle (Eq 25)
}
\details{
Return solar decimation in rad. Evaluates the equation:

w_s = acos(-tan(phi) * tan(delta))
}
\examples{
lat <- 13.1
date <- "2019-04-15"
w_s <- get_w_s_angle(lat, date)
}
