% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fao_penman_monteith.R
\name{read_et0_csv}
\alias{read_et0_csv}
\title{Read ET0 CSV}
\usage{
read_et0_csv(path)
}
\arguments{
\item{path}{Path to ET0 csv file}
}
\description{
Read ET0 CSV
}
\examples{
# Locate the path of the example et0 csv file
et0_path <- system.file("extdata", "ET0 - 20.59-78.96.csv", package = "fao56")

# Read the file
et0 <- read_et0_csv(et0_path)
}
